rclient: force /bin/sh on remote machine
Emacs will normally attempt to use the local shell path when trying to
run a shell on the remote system.  This is a problem when the local
shell doesn't exist on the remote system (for example,
/usr/local/bin/bash).  The solution is to just force remote requests
to be handled by /bin/sh, which is guaranteed to exist on Unix-based
systems.